Move to separate controller
