+++ app/cappy/controllers.rb
@@ -4,6 +4,7 @@ module Cappy
   module Controllers
     autoload :Base, 'cappy/controllers/base'
     autoload :Devices, 'cappy/controllers/devices'
+    autoload :States, 'cappy/controllers/states'
     autoload :Version, 'cappy/controllers/version'
   end
 end
