+++ app/cappy/controllers/devices.rb
@@ -36,18 +36,6 @@ module Cappy
         status 200
         Services::Devices.update(device_id, last_check_in: Time.now.utc)
       end
-
-      get '/devices/:device_id/state/?' do |device_id|
-        status 200
-        device = Services::Devices.get_device(device_id)
-        Services::States.read(device).to_json
-      end
-
-      post '/devices/:device_id/state/?' do |device_id|
-        status 201
-        device = Services::Devices.get_device(device_id)
-        Services::States.create(device, parse_json(req_body)).to_json
-      end
     end
   end
 end
